% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-act-extractiva.R
\name{plot_act_extractiva}
\alias{plot_act_extractiva}
\title{plot_act_extractiva}
\usage{
plot_act_extractiva(
  datos,
  col_especies,
  especie,
  col_caleta,
  orden_caletas,
  nombre_salida_prefix,
  ancho,
  alto
)
}
\arguments{
\item{datos}{dataframe de entrada. Formato: separado por tab. (Tiene que tener: coluna que identifique a las especies/recurso, caletas, meses, años. Los meses representan el desembarco en ese mes/año)}

\item{col_especies}{columna que tiene que el nombre de las especies/recursos}

\item{especie}{Vector con el nombre de la especie que se quiera incluir en el gráfico.}

\item{orden_caletas}{Vector con el nombre de las caletas que se quieran incluir en el gráfico en el orde que se quiera que aparezcan.}

\item{nombre_salida_prefix}{prefijo usado para el conjunto de gráficos a exportar}

\item{ancho}{ancho de la imágen}

\item{alto}{altura de la imágen}

\item{col_caletas}{columna que tiene que el nombre de las caletas que se quieran incluir en el gráfico}
}
\value{
imágenes png
}
\description{
Función para hacer el gráfico del "multipanel"
}
\examples{
\dontrun{
datos <- read_tsv("datos_act_extractiva.tsv", show_col_types = FALSE)
col_especie <- "Recurso"
especie <- "Sierra"
col_caleta <- "Localidad"
orden_caletas <- c("Mehuín", "Los Molinos", "Niebla", "Corral")
nombre_salida_prefix <- "plots_biobio_test"
ancho <- 8
alto <- 5
plot_act_extractiva(datos = datos, col_especie = col_especie, especie = especie, col_caleta = col_caleta, col_especie = col_especie, orden_caletas = orden_caletas, nombre_salida_prefix = nombre_salida_prefix, ancho = ancho, alto = alto)
}

}
