% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pesquerias.R
\name{plot_barra_pescadores}
\alias{plot_barra_pescadores}
\title{plot_barra_pescadores}
\usage{
plot_barra_pescadores(
  datos,
  caletas = NULL,
  col_caleta,
  ylab_text = "N° de Pescadores Inscritos",
  xlab_text = "Caletas",
  n_ticks,
  ancho,
  alto,
  nombre_salida
)
}
\arguments{
\item{datos}{dataframe de entrada. Formato: separado por tab (Tiene que tener las columnas: caletas/caleta_inscripcion, material,tipo)}

\item{caletas}{Vector con el nombre de las caletas que se quieran incluir en el gráfico. Si NULL, se usaran todas las presentes en el set  de datos}

\item{col_caleta}{columna que tiene que el nombre de las caletas}

\item{ylab_text}{Cadena de texto para el eje y del panel. Si NULL == "N° de Pescadores Inscritos"}

\item{xlab_text}{Cadena de texto para el eje x. Si NULL == "Caletas"}

\item{n_ticks}{Número de divisiones ejes y}

\item{ancho}{ancho de la imágen}

\item{alto}{altura de la imágen}

\item{nombre_salida}{Nombre figura incluyendo la extensión ("XXXX.png")}
}
\value{
imágenes png
}
\description{
Función para hacer el grafico de barra muestra el total de pescadores por caleta junto a la categoria de pescadores
}
\examples{
\dontrun{
datos <- read_tsv("data_pescador.txt", show_col_types = FALSE)
col_caleta <- "Caleta"
ylab
alto <- 6
ancho <- 9
n_ticks <- 6
nombre_salida <- "test_plot_barra_pescadores.png"
plot_barra_pescadores(datos = datos, col_caleta = col_caleta, alto = alto, ancho = ancho, n_ticks = n_ticks, nombre_salida = nombre_salida)
}
}
