% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_embarcaciones}
\alias{plot_embarcaciones}
\title{plot_tipo_embarcacion}
\usage{
plot_embarcaciones(
  datos,
  caletas = NULL,
  col_caleta,
  n_ticks = 6,
  alto,
  ancho,
  nombre_salida
)
}
\arguments{
\item{datos}{dataframe de entrada. Formato: separado por tab (Tiene que tener las columnas: caletas/caleta_inscripcion, material,tipo)}

\item{caletas}{Vector con el nombre de las caletas que se quieran incluir en el gráfico. Si NULL, se usaran todas}

\item{col_caleta}{columna que tiene que el nombre de las caletas}

\item{n_ticks}{Número de divisiones ejes y}

\item{alto}{altura de la imágen}

\item{ancho}{ancho de la imágen}

\item{nombre_salida}{Nombre figura incluyendo la extensión ("XXXX.png")}

\item{interv_potencia}{amplitud del intervalo para el subplot de potencia de las embarcaciones}
}
\value{
imágenes png
}
\description{
Función para hacer el grafico del "tipo de embarcaciones"
}
\examples{
\dontrun{
datos <- read_tsv("datos_tipo_embarcacion.txt", show_col_types = FALSE)
caletas <- "SAN VICENTE"
col_caleta <- "Caleta Inscripción"
n_ticks <- 6
interv_potencia <- 100
alto <- 6
ancho <- 8
nombre_salida <- "test_plot_embarcaciones.png"
plot_embarcaciones(datos = datos,caletas = caletas,col_caleta = col_caleta,n_ticks = n_ticks,alto = alto,ancho = ancho,nombre_salida = nombre_salida)
}
}
