% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_multipanel}
\alias{plot_multipanel}
\title{plot_multipanel}
\usage{
plot_multipanel(
  data,
  dicc,
  caletas = NULL,
  sp_remove = NULL,
  ylab_text_a,
  xlab_text_a,
  ylab_text_c,
  xlab_text_c,
  n_break_ylab,
  width,
  height,
  output_name,
  n_pie,
  round
)
}
\arguments{
\item{data}{dataframe de entrada}

\item{dir_input}{directory en donde se almacenan las imágenes L2}

\item{dir_output}{directory en donde se almacenaran las imágenes L3 resultantes}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{n_cores}{número de núcleos a usar. Por defecto, n_cores = 1 (corrida secuencial). La parelelización es respecto de la cantidad de sub_folder procesados simultaneamente}

\item{res_l2}{resolución para l2bin. Por defecto, res = "1" (H: 0.5km, Q: 250m, HQ: 100m, HH: 50m, 1: 1.1km, 2: 2.3km, 4: 4.6km, 9: 9.2km, 18: 18.5km, 36: 36km, 1D: 1 degree, HD: 0.5 degree, QD: 0.25 degree)}

\item{res_l3}{resolución para l3mapgen. Por defecto, res = "1km" (36km: 1080 x 540, 18km: 2160 x 1080, 9km: 4320 x 2160, 4km: 8640 x 4320, 2km: 17280 x 8640, 1km: 34560 x 17280, hkm: 69120 x 34560, qkm: 138240 x 69120, smi: 4096 x 2048, smi4: 8192 x 4096, land: 8640 x 4320)}

\item{north}{latitud norte para la generación de las imágenes L3}

\item{south}{latitud sur para la generación de las imágenes L3}

\item{west}{latitud oeste para la generación de las imágenes L3}

\item{east}{latitud este para la generación de las imágenes L3}

\item{need_extract}{mantener sistema de archivos año/mes? (TRUE/FALSE).Por defecto, FALSE}

\item{sort_files}{crearr sistema de archivos año/mes? (TRUE/FALSE).Por defecto, FALSE}
}
\value{
imágenes L3
}
\description{
Función para realizar gráfico multipanel
}
\examples{
\dontrun{
code here
}
}
