% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pesquerias.R
\name{plot_torta_pescadores}
\alias{plot_torta_pescadores}
\title{plot_torta_pescadores}
\usage{
plot_torta_pescadores(
  datos,
  col_caleta,
  caletas,
  ancho,
  alto,
  n_col,
  nombre_salida
)
}
\arguments{
\item{datos}{dataframe de entrada. Formato: separado por tab (Tiene que tener las columnas: caletas/caleta_inscripcion, material,tipo)}

\item{col_caleta}{columna que tiene que el nombre de las caletas}

\item{caletas}{Vector con el nombre de las caletas que se quieran incluir en el gráfico. Si NULL, se usaran todas las presentes en el set  de datos}

\item{ancho}{ancho de la imágen}

\item{alto}{altura de la imágen}

\item{nombre_salida}{Nombre figura incluyendo la extensión ("XXXX.png")}

\item{ncol}{numero de columnas de la imágen}
}
\value{
imágenes png
}
\description{
Función para hacer el grafico de torta que categorias de pescador por caleta
}
\examples{
\dontrun{
datos <- read_tsv("data_pescador.txt", show_col_types = FALSE)
col_caleta <- "Caleta"
caletas <- unique(datos$Caleta)[3:6]
alto <- 6
ancho <- 9
n_col <- 2
nombre_salida <- "test_plot_torta_caletas.png"
plot_torta_pescadores(datos = datos, caletas = caletas, col_caleta = col_caleta, alto = alto, ancho = ancho, n_col = n_col, nombre_salida = nombre_salida)
}
}
