% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pesquerias.R
\name{plot_multipanel}
\alias{plot_multipanel}
\title{plot_multipanel}
\usage{
plot_multipanel(
  datos,
  dicc,
  caletas = NULL,
  especies_rm = NULL,
  col_especie,
  ylab_text_sup,
  xlab_text_sup,
  ylab_text_inf,
  xlab_text_inf,
  ancho,
  alto,
  nombre_salida = "plot.png",
  n_ticks = 6,
  n_pie = 5,
  dec_red = 2
)
}
\arguments{
\item{datos}{dataframe de entrada. Formato: separado por tab. (Tiene que tener: coluna que identifique a las especies/recurso, caletas, meses, años. Los meses representan el desembarco en ese mes/año)}

\item{dicc}{archivo diccionario de recursos, separado por "," XXX}

\item{caletas}{Vector con el nombre de las caletas que se quieran incluir en el gráfico. Si NULL, se usaran todas}

\item{especies_rm}{Vector con el nombre de las especies que se quieran eliminar del gráfico en el gráfico. Si NULL, se usaran todas las especies}

\item{ylab_text_sup}{Cadena de texto para el eje y del panel superior. Si NULL == "Desembarque total (Ton)"}

\item{xlab_text_sup}{Cadena de texto para el eje x del panel superior. Si NULL == "Años"}

\item{ylab_text_inf}{Cadena de texto para el eje y del panel inferior. Si NULL == "Desembarque promedio (Ton)"}

\item{xlab_text_inf}{Cadena de texto para el eje x del panel inferior. Si NULL == "Meses"}

\item{ancho}{ancho de la imágen}

\item{alto}{altura de la imágen}

\item{nombre_salida}{Nombre figura incluyendo la extensión ("XXXX.png")}

\item{n_ticks}{Número de divisiones ejes y}

\item{n_pie}{Número de divisiones de la torta}

\item{dec_red}{Número de decimales para los porcentajes de los gráficos de torta}

\item{col_especies}{columna que tiene que el nombre de las especies/recursos}
}
\value{
imágenes png
}
\description{
Función para hacer el gráfico del "multipanel"
}
\examples{
\dontrun{
datos <- read_tsv("data.txt", show_col_types = FALSE)
dicc <- read_delim("diccionario_recursos.txt", delim = ";", show_col_types = FALSE)
caletas <- "SAN VICENTE"
col_especie <- "ESPECIE"
especies_rm <- "SARDINA COMUN"
ylab_text_sup <- "eje y"
xlab_text_sup <- "eje x"
ylab_text_inf <- "eje y"
xlab_text_inf <- "eje x"
ancho <- 12
alto <- 8
n_ticks <- 6
nombre_salida <- "test_plot_multipanel"
n_pie <- 5
dec_red <- 2
plot_multipanel(datos = datos, dicc = dicc, caletas = caletas, especies_rm = especies_rm, col_especie = col_especie, ylab_text_sup = ylab_text_sup, xlab_text_sup = xlab_text_sup, ylab_text_inf = ylab_text_inf, xlab_text_inf = xlab_text_inf, ancho = ancho, alto = alto, nombre_salida = nombre_salida, n_tick = n_tick, n_pie = n_pie, dec_red = dec_red)
}
}
